mercury-ai/mercury-bot/requirements.txt
fastapi
uvicorn
google-generativai
python-dotenv